+++ packages/embed/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @opencx/widget
 
+## 4.0.27
+
+### Patch Changes
+
+- fix loading state when awaiting bot reply
+
 ## 4.0.26
 
 ### Patch Changes