+++ packages/react-headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opencx/widget-react-headless
 
+## 4.0.27
+
+### Patch Changes
+
+- fix loading state when awaiting bot reply
+- Updated dependencies
+  - @opencx/widget-core@4.0.27
+
 ## 4.0.26
 
 ### Patch Changes